Divided view into separate files for maintainability and clarity
