+++ sd_food_bank_ai_bot/admin_panel/views/__init__.py
@@ -1,0 +1,2 @@
+from .admin_panel_faq import *
+from .phone_service_faq import *
